refactor: extract matrix conversation operations
